refactor: remove Transfer event fetching from free tier
- Transfers are usually waaaay more than Approval events, so removing
  them reduces load on infra by a lot
- Remove NFT balances as we cannot determine it without transfers
- Remove spam airdrop checks as we cannot determine it either
- Move balance fetching out of allowance derivation and into an async
  process similar to price data enrichment
- Small other refactors
- Note: this removes some tokens from the permit cancel dashboard; it
  should be decided later what to do about this

+++ apps/web/components/allowances/dashboard/cells/AssetCell.tsx
@@ -1,7 +1,9 @@
 'use client';
 
-import type { TokenData } from '@revoke.cash/core/tokens';
+import { isErc721Contract, type TokenData } from '@revoke.cash/core/tokens';
+import { isNullish } from '@revoke.cash/core/utils';
 import { formatBalance, formatFiatBalance } from '@revoke.cash/core/utils/formatting';
+import Loader from 'components/common/Loader';
 import AssetDisplay from './AssetDisplay';
 
 interface Props {
@@ -9,17 +11,32 @@ interface Props {
 }
 
 const AssetCell = ({ asset }: Props) => {
+  if (isErc721Contract(asset.contract)) {
+    return (
+      <div className="flex items-center gap-1 py-1 w-48 lg:w-56 h-12">
+        <AssetDisplay asset={asset} />
+      </div>
+    );
+  }
+
+  const isBalanceLoading = isNullish(asset.balance);
   const balanceText = formatBalance(asset.metadata.symbol, asset.balance, asset.metadata.decimals);
   const fiatBalanceText = formatFiatBalance(asset.balance, asset.metadata.price, asset.metadata.decimals);
 
   return (
-    <div className="flex items-center gap-1 py-1 w-48 lg:w-56">
+    <div className="flex items-center gap-1 py-1 w-48 lg:w-56 h-12">
       <div className="flex flex-col items-start gap-0.5">
         <AssetDisplay asset={asset} />
 
         <div className="text-xs text-zinc-500 dark:text-zinc-400 flex gap-1 w-48 lg:w-56">
-          <div className="truncate shrink">{balanceText}</div>
-          {fiatBalanceText ? <div className="grow shrink-0">({fiatBalanceText})</div> : null}
+          {isBalanceLoading ? (
+            <Loader isLoading className="h-4 w-24 rounded-sm" />
+          ) : (
+            <>
+              <div className="truncate shrink">{balanceText}</div>
+              {fiatBalanceText ? <div className="grow shrink-0">({fiatBalanceText})</div> : null}
+            </>
+          )}
         </div>
       </div>
     </div>
